Remove old structure
